+++ packages/config/src/config.js
@@ -5,6 +5,9 @@ import {
   SUBSCRIBE,
   UNSUBSCRIBE,
 } from "@onflow/util-actor"
+import * as logger from "@onflow/util-logger"
+import {invariant} from "@onflow/util-invariant"
+import {getContracts, cleanNetwork, anyHasPrivateKeys} from "../utils/utils"
 
 const NAME = "config"
 const PUT = "PUT_CONFIG"
@@ -61,15 +64,33 @@ const HANDLERS = {
 
 spawn(HANDLERS, NAME)
 
+/**
+ * @description Adds a key-value pair to the config
+ * @param {string} key - The key to add
+ * @param {*} value - The value to add
+ * @returns {Promise<object>} - The current config
+ */
 function put(key, value) {
   send(NAME, PUT, {key, value})
   return config()
 }
 
+/**
+ * @description Gets a key-value pair with a fallback from the config
+ * @param {string} key - The key to add
+ * @param {*} [fallback] - The fallback value to return if key is not found
+ * @returns {Promise<*>} - The value found at key or fallback
+ */
 function get(key, fallback) {
   return send(NAME, GET, {key, fallback}, {expectReply: true, timeout: 10})
 }
 
+/**
+ * @description Returns the first non null config value or the fallback
+ * @param {string[]} wants - The keys to search for
+ * @param {*} fallback - The fallback value to return if key is not found
+ * @returns {Promise<*>} - The value found at key or fallback
+ */
 async function first(wants = [], fallback) {
   if (!wants.length) return fallback
   const [head, ...rest] = wants
@@ -78,36 +99,138 @@ async function first(wants = [], fallback) {
   return ret
 }
 
+/**
+ * @description Returns the current config
+ * @returns {Promise<object>} - The current config
+ */
 function all() {
   return send(NAME, GET_ALL, null, {expectReply: true, timeout: 10})
 }
 
+/**
+ * @description Updates a key-value pair in the config
+ * @param {string} key - The key to update
+ * @param {Function} fn - The function to update the value with
+ * @returns {Promise<object>} - The current config
+ */
 function update(key, fn = identity) {
   send(NAME, UPDATE, {key, fn})
   return config()
 }
 
+/**
+ * @description Deletes a key-value pair from the config
+ * @param {string} key - The key to delete
+ * @returns {Promise<object>} - The current config
+ */
 function _delete(key) {
   send(NAME, DELETE, {key})
   return config()
 }
 
+/**
+ * @description Returns a subset of the config based on a pattern
+ * @param {string} pattern - The pattern to match keys against
+ * @returns {Promise<object>} - The subset of the config
+ */
 function where(pattern) {
   return send(NAME, WHERE, {pattern}, {expectReply: true, timeout: 10})
 }
 
+/**
+ * @description Subscribes to config updates
+ * @param {Function} callback - The callback to call when config is updated
+ * @returns {Function} - The unsubscribe function
+ */
 function subscribe(callback) {
   return subscriber(NAME, () => spawn(HANDLERS, NAME), callback)
 }
 
+/**
+ * @description Clears the config
+ * @returns {void}
+ */
 export function clearConfig() {
   return send(NAME, CLEAR)
 }
 
+/**
+ * @description Resets the config to a previous state
+ * @param {object} oldConfig - The previous config state
+ * @returns {Promise<object>} - The current config
+ */
 function resetConfig(oldConfig) {
   return clearConfig().then(config(oldConfig))
 }
 
+/**
+ * @description Takes in flow.json or array of flow.json files and creates contract placeholders
+ * @param {object|object[]} data - The flow.json or array of flow.json files
+ * @returns {void}
+ */
+async function load(data) {
+  const network = await get("flow.network")
+  const cleanedNetwork = cleanNetwork(network)
+  const {flowJSON} = data
+
+  invariant(Boolean(flowJSON), "config.load -- 'flowJSON' must be defined")
+
+  invariant(
+    cleanedNetwork,
+    `Flow Network Required -- In order for FCL to load your contracts please define "flow.network" to "emulator", "local", "testnet", or "mainnet" in your config. See more here: https://developers.flow.com/tools/fcl-js/reference/configure-fcl`
+  )
+
+  if (anyHasPrivateKeys(flowJSON)) {
+    const isEmulator = cleanedNetwork === "emulator"
+
+    logger.log({
+      title: "Private Keys Detected",
+      message: `Private keys should be stored in a separate flow.json file for security. See more here: https://developers.flow.com/tools/flow-cli/security`,
+      level: isEmulator ? logger.LEVELS.warn : logger.LEVELS.error,
+    })
+
+    if (!isEmulator) return
+  }
+
+  for (const [key, value] of Object.entries(
+    getContracts(flowJSON, cleanedNetwork)
+  )) {
+    const contractConfigKey = `0x${key}`
+    const existingContractConfigKey = await get(contractConfigKey)
+    if (existingContractConfigKey && existingContractConfigKey !== value) {
+      logger.log({
+        title: "Contract Placeholder Conflict Detected",
+        message: `A generated contract placeholder from config.load conflicts with a placeholder you've set manually in config have the same name.`,
+        level: logger.LEVELS.warn,
+      })
+    } else {
+      put(contractConfigKey, value)
+    }
+
+    const systemContractConfigKey = `system.contracts.${key}`
+    const systemExistingContractConfigKeyValue = await get(
+      systemContractConfigKey
+    )
+    if (
+      systemExistingContractConfigKeyValue &&
+      systemExistingContractConfigKeyValue !== value
+    ) {
+      logger.log({
+        title: "Contract Placeholder Conflict Detected",
+        message: `A generated contract placeholder from config.load conflicts with a placeholder you've set manually in config have the same name.`,
+        level: logger.LEVELS.warn,
+      })
+    } else {
+      put(systemContractConfigKey, value)
+    }
+  }
+}
+
+/**
+ * @description Sets the config
+ * @param {object} [values] - The values to set
+ * @returns {object} - The config methods
+ */
 function config(values) {
   if (values != null && typeof values === "object") {
     Object.keys(values).map(d => put(d, values[d]))
@@ -123,6 +246,7 @@ function config(values) {
     where,
     subscribe,
     overload,
+    load,
   }
 }
 
@@ -135,6 +259,7 @@ config.delete = _delete
 config.where = where
 config.subscribe = subscribe
 config.overload = overload
+config.load = load
 
 export {config}