+++ packages/fcl/src/default-config.js
@@ -1,4 +1,5 @@
 import {config} from "@onflow/config"
+import {setChainIdDefault} from "./utils/getChainId"
 
 const isServerSide = () => typeof window === "undefined"
 
@@ -12,6 +13,10 @@ config({
   "discovery.wallet.method.default": "IFRAME/RPC",
   "fcl.storage.default": SESSION_STORAGE,
 })
+
+// this is an async function but we can't await bc it's run at top level.
+// NOT guaranteed that flow.network.default is set after this call (or at startup)
+setChainIdDefault()
 
 export async function configLens(regex) {
   return Object.fromEntries(