+++ packages/fcl/src/interaction-template-utils/generate-dependency-pin.js
@@ -11,6 +11,15 @@ import {genHash} from "./utils/hash.js"
 import {findImports} from "./utils/find-imports.js"
 import {generateImport} from "./utils/generate-import.js"
 
+/**
+ * @description Produces a dependency pin for a contract at a given block height
+ * @param {object} params
+ * @param {string} params.address - The address of the account containing the contract
+ * @param {string} params.contractName - The name of the contract
+ * @param {number} params.blockHeight - The block height to produce the dependency pin for
+ * @param {object} opts - Options to pass to the interaction
+ * @returns {Promise<string>} - The dependency pin
+ */
 export async function generateDependencyPin(
   {address, contractName, blockHeight},
   opts = {}
@@ -72,6 +81,14 @@ export async function generateDependencyPin(
   return genHash(contractHashesJoined)
 }
 
+/**
+ * @description Produces a dependency pin for a contract at latest sealed block
+ * @param {object} params
+ * @param {string} params.address - The address of the account containing the contract
+ * @param {string} params.contractName - The name of the contract
+ * @param {object} opts - Options to pass to the interaction
+ * @returns {Promise<string>} - The dependency pin
+ */
 export async function generateDependencyPinAtLatestSealedBlock(
   {address, contractName},
   opts = {}