+++ packages/fcl/src/interaction-template-utils/get-interaction-template-audits.js
@@ -3,7 +3,18 @@ import {log, LEVELS} from "@onflow/util-logger"
 import {query} from "../exec/query.js"
 import {generateTemplateId} from "./generate-template-id.js"
 import {normalizeInteractionTemplate} from "../normalizers/interaction-template/interaction-template.js"
+import {getChainId} from "../utils"
 
+/**
+ * @description Returns whether a set of auditors have audited a given Interaction Template
+ * 
+ * @param {object} params
+ * @param {object} params.template - Interaction Template
+ * @param {Array<string>} params.auditors - Array of auditors
+ * @param {object} opts
+ * @param {string} opts.flowInteractionAuditContract - Flow Interaction Template Audit contract address
+ * @returns {Promise<object>} - Object of auditor addresses and audit status
+ */
 export async function getInteractionTemplateAudits(
   {template, auditors},
   opts = {}
@@ -52,7 +63,7 @@ export async function getInteractionTemplateAudits(
 
       let FlowInteractionAuditContract = opts.flowInteractionAuditContract
       if (!FlowInteractionAuditContract) {
-        const fclNetwork = await config().get("flow.network")
+        const fclNetwork = await getChainId()
         invariant(
           fclNetwork === "mainnet" || fclNetwork === "testnet",
           "getInteractionTemplateAudits Error: Unable to determine address for FlowInteractionTemplateAudit contract. Set configuration for 'fcl.network' to 'mainnet' or 'testnet'"