+++ packages/sdk/src/decode/decode.js
@@ -133,6 +133,13 @@ const recurseDecode = async (decodeInstructions, decoders, stack) => {
   return await decoder(decodeInstructions.value, decoders, stack)
 }
 
+/**
+ * @description - Decodes a response from Flow into JSON
+ * @param {*} decodeInstructions - The response object from Flow
+ * @param {object} customDecoders - An object of custom decoders
+ * @param {Array<*>} stack - The stack of the current decoding
+ * @returns {Promise<*>} - The decoded response
+ */
 export const decode = async (
   decodeInstructions,
   customDecoders = {},
@@ -206,6 +213,16 @@ export const decodeResponse = async (response, customDecoders = {}) => {
     return response.transactionId
   } else if (response.collection) {
     return response.collection
+  } else if (response.networkParameters) {
+    const chainIdMap = {
+      "flow-testnet": "testnet",
+      "flow-mainnet": "mainnet",
+      "flow-emulator": "local",
+    }
+
+    return {
+      chainId: chainIdMap[response.networkParameters.chainId],
+    }
   }
 
   return null