+++ packages/util-logger/src/util-logger.js
@@ -1,5 +1,16 @@
 import {config} from "@onflow/config"
 
+/**
+ * The levels of the logger
+ * 
+ * @typedef {Object} LEVELS
+ * @property {number} debug - The debug level
+ * @property {number} info - The info level
+ * @property {number} log - The log level
+ * @property {number} warn - The warn level
+ * @property {number} error - The error level
+ * 
+ */
 export const LEVELS = Object.freeze({
   debug: 5,
   info: 4,
@@ -8,6 +19,17 @@ export const LEVELS = Object.freeze({
   error: 1,
 })
 
+/**
+ * Builds a message formatted for the logger
+ * 
+ * @param {Object} options - The options for the log
+ * @param {string} options.title - The title of the log
+ * @param {string} options.message - The message of the log
+ * @returns {Array<string>} - The message formatted for the logger
+ * 
+ * @example
+ * buildLoggerMessageArgs({ title: "My Title", message: "My Message" })
+ */
 const buildLoggerMessageArgs = ({title, message}) => {
   return [
     `
@@ -25,6 +47,20 @@ const buildLoggerMessageArgs = ({title, message}) => {
   ]
 }
 
+/**
+ * Logs messages based on the level of the message and the level set in the config
+ * 
+ * @param {Object} options - The options for the log
+ * @param {string} options.title - The title of the log
+ * @param {string} options.message - The message of the log
+ * @param {number} options.level - The level of the log
+ * @param {boolean} options.always - Whether to always show the log
+ * @returns {Promise<void>}
+ * 
+ * @example
+ * log({ title: "My Title", message: "My Message", level: LEVELS.warn, always: false })
+ * 
+ */
 export const log = async ({title, message, level, always = false}) => {
   const configLoggerLevel = await config.get("logger.level", LEVELS.warn)
 
@@ -51,6 +87,22 @@ export const log = async ({title, message, level, always = false}) => {
   }
 }
 
+/**
+ * Logs a deprecation notice
+ * 
+ * @param {Object} options - The options for the log
+ * @param {string} options.pkg - The package that is being deprecated
+ * @param {string} options.subject - The subject of the deprecation
+ * @param {string} options.transition - The transition path for the deprecation
+ * @param {number} options.level - The level of the log
+ * @param {string} options.message - The message of the log
+ * @param {Function} options.callback - A callback to run after the log
+ * @returns {Promise<void>}
+ * 
+ * @example
+ * log.deprecate({ pkg: "@onflow/fcl", subject: "Some item", transition: "https://github.com/onflow/flow-js-sdk", message: "Descriptive message", level: LEVELS.warn, callback: () => {} })
+ * 
+ */
 log.deprecate = ({
   pkg,
   subject,