+++ packages/transport-grpc/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @onflow/transport-grpc
 
+## 1.2.0-alpha.1
+
+### Minor Changes
+
+- [#1577](https://github.com/onflow/fcl-js/pull/1577) [`d9a49531`](https://github.com/onflow/fcl-js/commit/d9a495316cd03ed0de99e0f01d1b8850a1f0eec4) Thanks [@chasefleming](https://github.com/chasefleming)! - Add npmignore file for build
+
+### Patch Changes
+
+- Updated dependencies [[`d9a49531`](https://github.com/onflow/fcl-js/commit/d9a495316cd03ed0de99e0f01d1b8850a1f0eec4)]:
+  - @onflow/protobuf@1.2.0-alpha.0
+  - @onflow/rlp@1.1.0-alpha.0
+  - @onflow/util-address@1.1.0-alpha.3
+  - @onflow/util-invariant@1.1.0-alpha.0
+  - @onflow/util-template@1.1.0-alpha.0
+
+## 1.2.0-alpha.0
+
+### Minor Changes
+
+- [#1420](https://github.com/onflow/fcl-js/pull/1420) [`c20bc344`](https://github.com/onflow/fcl-js/commit/c20bc34448a22966d349b0b1c4c0f742ae93a355) Thanks [@huyndo](https://github.com/huyndo)! - Add GetNetworkParameters interaction and a util to get chain ID
+
+### Patch Changes
+
+- Updated dependencies [[`5bec5576`](https://github.com/onflow/fcl-js/commit/5bec5576a79809d0684411736e3f4c02b8051c22)]:
+  - @onflow/util-address@1.1.0-alpha.0
+
+## 1.1.4
+
+### Patch Changes
+
+- [#1476](https://github.com/onflow/fcl-js/pull/1476) [`458e2914`](https://github.com/onflow/fcl-js/commit/458e291416fcd6237b4dfd82cc558ec0a0ef4d03) Thanks [@justinbarry](https://github.com/justinbarry)! - Remove logic to retry 503/504 http status responses
+
 ## 1.1.3
 
 ### Patch Changes