+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @onflow/types
 
+## 1.1.0-alpha.0
+
+### Minor Changes
+
+- [#1577](https://github.com/onflow/fcl-js/pull/1577) [`d9a49531`](https://github.com/onflow/fcl-js/commit/d9a495316cd03ed0de99e0f01d1b8850a1f0eec4) Thanks [@chasefleming](https://github.com/chasefleming)! - Add npmignore file for build
+
 ## 1.0.5
 
 ### Patch Changes